+++ tests/test_opaque_types.py
@@ -4,21 +4,21 @@ from pybind11_tests import ConstructorStats, UserType
 
 
 def test_string_list():
-    l = m.StringList()
-    l.push_back("Element 1")
-    l.push_back("Element 2")
-    assert m.print_opaque_list(l) == "Opaque list: [Element 1, Element 2]"
-    assert l.back() == "Element 2"
+    lst = m.StringList()
+    lst.push_back("Element 1")
+    lst.push_back("Element 2")
+    assert m.print_opaque_list(lst) == "Opaque list: [Element 1, Element 2]"
+    assert lst.back() == "Element 2"
 
-    for i, k in enumerate(l, start=1):
+    for i, k in enumerate(lst, start=1):
         assert k == "Element {}".format(i)
-    l.pop_back()
-    assert m.print_opaque_list(l) == "Opaque list: [Element 1]"
+    lst.pop_back()
+    assert m.print_opaque_list(lst) == "Opaque list: [Element 1]"
 
     cvp = m.ClassWithSTLVecProperty()
     assert m.print_opaque_list(cvp.stringList) == "Opaque list: []"
 
-    cvp.stringList = l
+    cvp.stringList = lst
     cvp.stringList.push_back("Element 3")
     assert m.print_opaque_list(cvp.stringList) == "Opaque list: [Element 1, Element 3]"