+++ tests/test_sequences_and_iterators.py
@@ -131,9 +131,9 @@ def test_python_iterator_in_cpp():
         m.iterator_to_list(iter(bad_next_call, None))
     assert str(excinfo.value) == "py::iterator::advance() should propagate errors"
 
-    l = [1, None, 0, None]
-    assert m.count_none(l) == 2
-    assert m.find_none(l) is True
+    lst = [1, None, 0, None]
+    assert m.count_none(lst) == 2
+    assert m.find_none(lst) is True
     assert m.count_nonzeros({"a": 0, "b": 1, "c": 2}) == 2
 
     r = range(5)